chore: move calendar markup to helpers.py
